requirements.txt
amqp                     
arabic-reshaper            
asgiref                    
asn1crypto                
async-timeout             
beautifulsoup4           
billiard                 
cachetools                
celery                    
certifi                   
cffi                      
chardet                   
charset-normalizer         
click                      
click-didyoumean           
click-plugins              
click-repl                
colorama                 
cron-descriptor            
cryptography               
cssselect                 
defusedxml                 
distlib                    
Django                     
django-allauth             
django-bootstrap-v5        
django-cache               
django-categories          
django-celery-beat        
django-channels           
django-ckeditor           
django-crispy-forms        
django-csp                
django-filter             
django-gunicorn            
django-haystack           
django-js-asset           
django-mptt                
django-pwa                 
django-redis               
django-rest-framework      
django-secure              
django-social-share        
django-theme-soft-design   
django-timezone-field      
django-versatileimagefield 
django-widget-tweaks       
djangorestframework        
elastic-transport          
elasticsearch              
filelock                   
gunicorn                   
h11                        
html5lib                   
idna                       
importlib                  
kombu                      
lxml                       
oauthlib                   
oscrypto                   
packaging                  
Pillow                     
pip                        
platformdirs               
pluggy                     
prompt-toolkit             
psycopg                   
psycopg-binary            
pycparser                  
pyHanko                    
pyhanko-certvalidator    
PyJWT                     
pypdf                      
pypng                     
python-bidi               
python-crontab           
python-dateutil          
python-magic              
python3-openid             
pytz                       
PyYAML                     
qrcode                   
redis                    
reportlab                  
requests                   
requests-oauthlib          
setuptools                 
six                       
soupsieve                  
sqlparse                   
svglib                   
tinycss                 
tomli                      
tox                        
typing_extensions          
tzdata                     
tzlocal                    
unicode-slugify            
Unidecode                  
uritools                   
urllib3                    
utils                     
uvicorn                    
vine                       
virtualenv                
wcwidth                    
webencodings             
whitenoise                
xhtml2pdf                  
